Move XdsChannelFactory definition into XdsClient file.
